update doanh thu

+++ src/main/java/com/group11/service/IRevenueService.java
@@ -3,12 +3,12 @@ package com.group11.service;
 import java.time.LocalDateTime;
 
 public interface IRevenueService {
-    public long calculateDailyRevenue();
+    public double calculateDailyRevenue();
     public long calculateWeeklyRevenue();
-    public long calculateMonthlyRevenue();
-    public long calculateYearlyRevenue();
+    public double calculateMonthlyRevenue();
+    public double calculateYearlyRevenue();
     public long calculateTotalRevenue();
-    public long calculateRevenue(LocalDateTime startDate, LocalDateTime endDate);
+    public double calculateRevenue(LocalDateTime startDate, LocalDateTime endDate);
     public double[] calculateMonthlyRevenueForYear();
     //public List<OrderEntity> getOrdersWithShippingStatusAndReceiveDate(LocalDateTime startDate, LocalDateTime endDate);
 }
